+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.02.19
+
+### Added
+
+- Designed a proof of concept for FPLDataTracker component
+
+### Updated
+
+- Changed design to include kernal and secondary methods. Also a main method to test.
+- Kernal Methods:
+  - addPlayer, removePlayer, updatePlayerStats, containsPlayer, getPlayerCount.
+- Secondary Methods:
+  - getPlayerStats, getTopScorers, getTopAssisters, getBestPlayerByPosition.
+
 ## 2025.02.03
 
 ### Added